Various changes run on Azure and VS 2022
The main web project had to be deleted, created anew, and then all of the individual files re-added to the project because it would not run with anything other than VS 2010.  It was incompatible and could not be automatically converted, and therefore, had to be recreated.

The DB and Reports projects had to be upgraded in order to be used with VS 2022.

For Azure:
* References to the email host needed to be update for use with smtp.sparkpostmail.com, and it's associated password.
* Report server BLL needed to be updated to include Report server credentials to allow on-prem generated reports to be displayed with Azure apps.
* Lastly, a number of supporting files, i.e. DLLs were added to the bin directory to allow this app to run due to its age.